refactor(counter): switch input type to text + state type to string
I really wanted to keep the input as a type number, but the input is a
complete wildcard. It kinda just does whatever it feels like, because of
the included input sanitization. There is no way to turn off this.

So I cut my loses and replaced it with a type text. I've also changed
the state to a string, since parsing it as a number will not keep a
trailing decimal number.