fix(servings): change where clause to include min and max
In an ideal world, the "date" field wouldn't have time. Unfortunately,
prisma does not support the type DATE on SQLite. I could set the "date"
field as a string, but for validation purposes I feel it's better to
leave it as a DATETIME for now.

To make sure when selecting it includes the right things, no matter if a
time is set, I'm setting a min date with the time 00:00 and a max date
with the time 23:59. This should avoid rows not being selected due to
whatever reason having a time set.