fix(counter): rerender when date changes
There was a bug when the date changed. When the initialValue was ''
(empty string) and the quantity had been updated, the counter display
value would not update. This would happen if, in the following date, the
initialValue was also '' (empty string). React would not rerender
because, technically, initialValue did not change even though the
quantity had been updated.