test(counter): move input specific tests to CounterInput test file
Since the Counter input has now network requests, I thought it would be
better architecturally to separate input features tests to its own file.

This came with its own challenges – CounterInput does not hold its own
state. It would have been possible to simply monitor calls to a mock
function, and rerender, however this proved to turn testing complicated
and annoying. Typing was especially problem, because the input rerenders
everytime a letter is typed. In the end it was causing too many problems
and limiting the types of tests I could run.

To solve this, I created a HOC (withState) that hold the state and
passes it as a prop to the child component. I though this would be a
useful component, so I made it as flexible as possible. Namely, the prop
name that holds the state can be provided as a prop to the HOC (e.g.
some child components may have "value" as the prop name, while other may
call it "state").